requirements.txt
asgiref==3.5.2
contourpy==1.0.6
cycler==0.11.0
Django==4.1.3
fonttools==4.38.0
kiwisolver==1.4.4
matplotlib==3.6.2
numpy==1.23.4
packaging==21.3
Pillow==9.3.0
pyparsing==3.0.9
python-dateutil==2.8.2
six==1.16.0
sqlparse==0.4.3
tzdata==2022.6
whitenoise==6.2.0
